REFACTOR: renamed modules & moved `RolesEnum` to logical place
