+++ src/config/index.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from "fs";
-import * as authorization from "./authorization";
+import * as authorization from "../common/RolesEnum";
 import logger from "./logger/index";
 import mongoose from "./mongoose/index";
 import passport from "./passport/index";
