+++ src/controllers/api/ContractTransactionsController.ts
@@ -1,5 +1,5 @@
-import { ShapesEnum } from "@/common/modelShaping";
-import { RolesEnum } from "@/config/authorization";
+import { RolesEnum } from "@/common/RolesEnum";
+import { ShapesEnum } from "@/common/ShapesEnum";
 import { AuthorizedRoles } from "@/middlewares/AuthorizedRoles";
 import { BaseContractTransactionModel } from "@/models/BaseContractTransactionModel";
 import { CreditTransactionModel } from "@/models/CreditTransactionModel";
