+++ src/models/UserModel.ts
@@ -1,5 +1,5 @@
-import { ShapesEnum } from "@/common/modelShaping";
-import { RolesEnum } from "@/config/authorization";
+import { RolesEnum } from "@/common/RolesEnum";
+import { ShapesEnum } from "@/common/ShapesEnum";
 import { CredentialsModel } from "@/models/CredentialsModel";
 import { Model, ObjectID, PostHook, PreHook, Sparse, Trim, Unique } from "@tsed/mongoose";
 import {
